String bpmap

+++ storage/FilePage.cpp
@@ -128,7 +128,7 @@ namespace bptreedb
 
 	void CFilePage::WriteCRC(byte_t* pData, size_t nSize)
 	{
-		uint32_t calcCrc = utils::Crc32(pData + sizeof(uint32_t), nSize - sizeof(uint32_t));
+		uint32_t calcCrc = (uint32_t)utils::Crc32(pData + sizeof(uint32_t), nSize - sizeof(uint32_t));
 		memcpy(pData, &calcCrc, sizeof(uint32_t));
 	}
 
@@ -136,7 +136,7 @@ namespace bptreedb
 	{
 		uint32_t crc = 0;
 		memcpy(&crc, pData, sizeof(uint32_t));
-		uint32_t calcCrc = utils::Crc32(pData + sizeof(uint32_t), nSize - sizeof(uint32_t));
+		uint32_t calcCrc = (uint32_t)utils::Crc32(pData + sizeof(uint32_t), nSize - sizeof(uint32_t));
 
 		return crc == calcCrc;
 	}
